+++ src/ruleset/rule-calculator.rb
@@ -54,29 +54,30 @@ exoplanet = World.new(160, 40000, 6356)
 head = Unit.new("Head", nil, 14, 120)
 
 os << "RESULTS"
-os << "_"
 
 os << "DD 4.1.1. - Time-limit/moves required for equatorial sectors"
 sectors = spec[:sector_grid]*18
 sector_side_km = exoplanet[:circumference_km]/spec[:sector_grid]
 sector_visit_min = sectors * limits[:sectors_visited_min]
-travel_min_km = sector_side_km * sector_visit_min 
+travel_min_km = sector_side_km * sector_visit_min
 travel_time_min_d = travel_min_km / (head[:work_hours] * head[:velocity_kmh])
 travel_time_min_turns = (travel_time_min_d/spec[:time_days_per_turn]).to_int
 os << "Head-unit time to travel #{sector_visit_min.to_int} of #{sectors} sectors (no elevation/lati multi):"
-os << "#{travel_time_min_d.to_int} days, #{travel_time_min_turns} turns, #{(spec[:time_turn_len_s]*travel_time_min_turns)/60} minutes."
+os << "In-game #{travel_time_min_d.to_int} days / #{travel_time_min_turns} turns, realtime #{(spec[:time_turn_len_s]*travel_time_min_turns)/60} minutes."
 
-os << "DD 4.1.1. - Polar latitudes cost: declination cost"
-os << "On a globe, sectors physical size decreases proportional to distance from equator. Optimally travel time for every #{360/spec[:sector_grid]} degree sector is similar, explained by harsher environments."
+os << "DD 4.1.1. - Polar movement cost, ie. declination"
+os << "On a globe, sector physical size decreases proportional to distance from equator. Optimally travel time for every #{360/spec[:sector_grid]} degree sector is similar, explained by harsher environments."
 sector_side_polar_km = exoplanet[:polar_radius]/spec[:sector_grid]
-os << "On a globe sector side in km: #{sector_side_km} equatorial and polar #{sector_side_polar_km}"
-os << "TODO"
-os << "_"
+os << "Globe sector side in km: #{sector_side_km} equatorial and polar #{sector_side_polar_km}."
 
 os << "DD 4.9. - Temporal Mechanics"
-turns_min = spec[:time_max_s]/spec[:time_turn_len_s]
-os << "For a #{spec[:time_max_s]/60} minute game, #{turns_min*spec[:time_days_per_turn]} in-world days." 
-os << "Minimum number of turns #{turns_min} for limits."
+turns_max = spec[:time_max_s]/spec[:time_turn_len_s]
+os << "For a #{spec[:time_max_s]/60} minute game, #{turns_max*spec[:time_days_per_turn]} in-world days." 
+os << "Maximum number of turns #{turns_max}."
+
+throw "RulesetLimitsError - head-unit travel lim @ limits[:sectors_visited_min]." unless travel_time_min_turns < turns_max 
+os << "✓ DD 4.1.1 & DD 4.9 PASS." 
+os << "_"
 
 #pseudo game logic to mock resultant gameplay
 #f.ex. mock_gameplay(buildingRandomizer)